Add some initial component structure for main page
Starting to scaffold the basic project structure for the application.
Components will go in the components folder with a shared subfolder for
layout / shared components. Underneath here components can be grouped by
page hierarchy.

Styles will be put next to the component they affect if possible or in
the styles folder if they are shared across multiple components.

We do need to figure out how to handle navigation across the app.
