+++ components/shared/header_styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from 'react-native'
+
+const styles = StyleSheet.create({
+  header: {
+    height: 50,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  headerText: {
+    fontSize: 28,
+  },
+})
+
+export default styles
